Added Renato's suggestions
