+++ src/qibo/quantum_info/__init__.py
@@ -1,5 +1,4 @@
 from qibo.quantum_info.basis import *
-from qibo.quantum_info.expressibility import *
 from qibo.quantum_info.metrics import *
 from qibo.quantum_info.random_ensembles import *
 from qibo.quantum_info.superoperator_transformations import *
